Move ScreenX to game
