+++ src/game/cGame_logic.cpp
@@ -49,9 +49,9 @@
 #include "sidebar/cSideBarFactory.h"
 #include "utils/RNG.hpp"
 #include "utils/cLog.h"
-#include "utils/cPlatformLayerInit.h"
+#include "game/cPlatformLayerInit.h"
 #include "utils/cSoundPlayer.h"
-#include "utils/cScreenInit.h"
+#include "game/cScreenInit.h"
 #include "utils/d2tm_math.h"
 #include "map/cPreviewMaps.h"
 #include "map/MapGeometry.hpp"
